// app/interface.ts
export interface simplifiedProduct{
    productName: string;
    price: number;
    inventory: number;
    image: string;
    category: string;
    slug: string;

}
export interface fullProduct{
    productName: string;
    price: number;
    inventory: number;
    image: string;
    description: string;
    slug: string;
    category: string;
    price_id: string;

}